refactor: BarMenu 경로 수정
